Cleaned up plumbing.
