+++ packages/devtools_app/test/chart_test.dart
@@ -2,7 +2,7 @@ import 'package:devtools_app/src/charts/chart_controller.dart';
 import 'package:devtools_app/src/charts/chart_trace.dart';
 import 'package:devtools_app/src/charts/chart.dart';
 import 'package:devtools_app/src/globals.dart';
-import 'package:devtools_app/src/memory/memory_screen.dart';
+import 'package:devtools_app/src/memory/memory_charts.dart';
 import 'package:devtools_app/src/service_manager.dart';
 import 'package:devtools_app/src/utils.dart';
 import 'package:devtools_shared/devtools_shared.dart';
@@ -58,7 +58,7 @@ void main() {
             symbol: ChartSymbol.disc,
             diameter: 1.5,
           ),
-          name: MemoryScreen.externalDisplay,
+          name: externalDisplay,
         );
 
         // Used Heap
@@ -69,7 +69,7 @@ void main() {
             symbol: ChartSymbol.disc,
             diameter: 1.5,
           ),
-          name: MemoryScreen.usedDisplay,
+          name: usedDisplay,
         );
 
         // Heap Capacity
@@ -80,7 +80,7 @@ void main() {
             diameter: 0.0,
             symbol: ChartSymbol.dashedLine,
           ),
-          name: MemoryScreen.allocatedDisplay,
+          name: allocatedDisplay,
         );
 
         // RSS
@@ -91,7 +91,7 @@ void main() {
             symbol: ChartSymbol.dashedLine,
             strokeWidth: 2,
           ),
-          name: MemoryScreen.rssDisplay,
+          name: rssDisplay,
         );
 
         expect(controller.traces.length, equals(4));
@@ -100,20 +100,20 @@ void main() {
             case 0:
               expect(externalTraceIndex, equals(0));
               expect(
-                  controller.traces[index].name, MemoryScreen.externalDisplay);
+                  controller.traces[index].name, externalDisplay);
               break;
             case 1:
               expect(usedTraceIndex, equals(1));
-              expect(controller.traces[index].name, MemoryScreen.usedDisplay);
+              expect(controller.traces[index].name, usedDisplay);
               break;
             case 2:
               expect(capacityTraceIndex, equals(2));
               expect(
-                  controller.traces[index].name, MemoryScreen.allocatedDisplay);
+                  controller.traces[index].name, allocatedDisplay);
               break;
             case 3:
               expect(rssTraceIndex, equals(3));
-              expect(controller.traces[index].name, MemoryScreen.rssDisplay);
+              expect(controller.traces[index].name, rssDisplay);
               break;
           }
         }
